Rename packages to com.commercetools
